Add files via upload
I had some trouble working with the environment that we had setup last time but either way I made some progress ...

Basically I wasn't able to edit and test my code using the Lib. config. so I created a new Project and started from scratch ...
Anyways ... I managed to implement some of the things we had mentioned on the call ...

See the examples in Program.fs ... we can now "safely" retrieve Dividends data as we simply output the raw data from Json (only for dividends)
Queries can be "built" using the BuildQueries module and to "download" the queries you have to call one single function -> getSeries .... which returns a QueryResponse which has 2 fields ... (1) Data and (2) ErrorLog ... (Not sure about the names, but this was the idea)

Also, I did a small refacto. of the pupulateQuotes function, which is now populateSeries.

Lastly, I discovered something nice about the JsonTypeProvider (you might already know this) ... so if a particular Json has 20 fields and you only want 5 of them ... just delete the unwanted fields from the raw string Json sample and voilà!